% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{addColumnNameForLabels}
\alias{addColumnNameForLabels}
\alias{addColumnNameForLabels.default}
\alias{addColumnNameForLabels.list}
\alias{addColumnNameForLabels.sf}
\alias{addColumnNameForLabels.SpatialPolygonsDataFrame}
\alias{addColumnNameForLabels.SpatVector}
\title{Add \code{shinyLabel} column (attribute) to spatial vectors}
\usage{
addColumnNameForLabels(x, columnNameForLabels, ...)

\method{addColumnNameForLabels}{default}(x, columnNameForLabels, ...)

\method{addColumnNameForLabels}{list}(x, columnNameForLabels, ...)

\method{addColumnNameForLabels}{sf}(x, columnNameForLabels, ...)

\method{addColumnNameForLabels}{SpatialPolygonsDataFrame}(x, columnNameForLabels, ...)

\method{addColumnNameForLabels}{SpatVector}(x, columnNameForLabels, ...)
}
\arguments{
\item{x}{a spatial vector object (\code{sf}, \code{SpatialPolygons}, \code{SpatVector})}

\item{columnNameForLabels}{character or integer identifying an existing column (attribute)
to use as the \code{shinyLabel}.}

\item{...}{Additional arguments passed to other methods (not used)}
}
\value{
a modified object with the same class as \code{x}
}
\description{
Add \code{shinyLabel} column (attribute) to spatial vectors
}
