% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{methods}
\name{area,map-method}
\alias{area,map-method}
\title{Calculate area of (named) objects the \code{map} object}
\usage{
\S4method{area}{map}(x)
}
\description{
Calculate area of (named) objects the \code{map} object
}
\seealso{
Other mapMethods: \code{\link{crs,map-method}},
  \code{\link{mapRm}}, \code{\link{rasters}},
  \code{\link{studyAreaName}}, \code{\link{studyArea}}
}
\concept{mapMethods}
