% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS.R
\name{areaAndPolyValue}
\alias{areaAndPolyValue}
\alias{areaAndPolyValue.Raster}
\alias{areaAndPolyValue.SpatRaster}
\title{Determine the area of each zone in a raster}
\usage{
areaAndPolyValue(ras, ...)

\method{areaAndPolyValue}{Raster}(ras, ...)

\method{areaAndPolyValue}{SpatRaster}(ras, ...)
}
\arguments{
\item{ras}{A \code{Raster} or \code{SpatRaster} object}

\item{...}{Additional arguments (not used)}
}
\value{
list containing: \code{sizeInHa}, the area; and \code{polyID}, the polygon ID.
}
\description{
Polygonize a raster and calculate the area of each polygon.
}
