% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_data.R
\name{assign_colours_to_df}
\alias{assign_colours_to_df}
\title{Assign colours to a dataframe for specified discrete columns.}
\usage{
assign_colours_to_df(
  dataframe.df,
  columns,
  my_palette = NULL,
  auto_assign = T,
  my_default_palette = NULL
)
}
\description{
Assign colours to a dataframe for specified discrete columns.
}
