% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{assign_shapes_to_df}
\alias{assign_shapes_to_df}
\title{Assign shapes to a dataframe for specified discrete columns.}
\usage{
assign_shapes_to_df(
  dataframe.df,
  columns,
  use_shapes = T,
  default_shape_type = "filled",
  shape_pool = NULL
)
}
\arguments{
\item{dataframe.df}{input dataframe}

\item{columns}{columns to assign shapes}

\item{use_shapes}{Boolean indicating whether to use multiple shapes (TRUE) or single shape (FALSE), e.g. 21. Default (TRUE)}

\item{shape_pool}{vector of shapes indices to choose from. Default = NULL}
}
\description{
Assign shapes to a dataframe for specified discrete columns.
}
