% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{buildMetadata}
\alias{buildMetadata}
\title{Build \code{map} object metadata table}
\usage{
buildMetadata(
  metadata,
  isStudyArea,
  isRasterToMatch,
  layerName,
  obj,
  columnNameForLabels,
  objHash,
  leaflet,
  envir,
  ...
)
}
\arguments{
\item{metadata}{metadata \code{data.table}.}

\item{isStudyArea}{Logical indicating whether \code{obj} a study area.}

\item{isRasterToMatch}{Logical indicating whether \code{obj} is a \code{rasterToMatch} layer.
If \code{TRUE}, then this layer can be accessed by \code{rasterToMatch(map)}.}

\item{layerName}{Character specifying a label for this layer.}

\item{obj}{The geospatial object being added to the \code{map} object.}

\item{columnNameForLabels}{Character specifying the column name to use for labels.}

\item{objHash}{Character specifying the digest hash of the object.}

\item{leaflet}{Logical or Character vector of path(s) to write tiles.
If \code{TRUE} or a character vector, then this layer will be added to a leaflet map.
For \code{RasterLayer} or \code{SpatRaster} object, this will make tiles.
If path is not specified, it will be the current path.
The tile base file path will be \code{paste0(layerName, "_", rndstr(1, 6))}.}

\item{envir}{TODO: description needed}

\item{...}{Additional arguments.}
}
\description{
Build \code{map} object metadata table
}
