% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{calculate_correlation_matrix}
\alias{calculate_correlation_matrix}
\title{Generate correlation matrix}
\usage{
calculate_correlation_matrix(mydata.m, method = "pearson", adjust = "BH")
}
\description{
Generate correlation matrix
}
