% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_data.R
\name{colour_list_from_dataframe}
\alias{colour_list_from_dataframe}
\title{Return list of colours for variable values from metadata dataframe
assumes _colour column present}
\usage{
colour_list_from_dataframe(dataframe.df, variable)
}
\arguments{
\item{variable}{column (variable) to get colours for}
}
\description{
Return list of colours for variable values from metadata dataframe
assumes _colour column present
}
