% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{\%<>\%}
\alias{\%<>\%}
\title{Assignment pipe}
\arguments{
\item{lhs}{An object which serves both as the initial value and as target.}

\item{rhs}{a function call using the magrittr semantics.}
}
\value{
None, used to update the value of lhs.
}
\description{
Pipe an object forward into a function or call expression and update the
`lhs` object with the resulting value.
Magrittr imported function, see details and examples in the magrittr package.
}
