% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{methods}
\name{crs,map-method}
\alias{crs,map-method}
\title{Extract the crs of a \code{map}}
\usage{
\S4method{crs}{map}(x, ...)
}
\arguments{
\item{x}{Raster* or Spatial object }

\item{...}{additional arguments (not implemented)}
}
\description{
Extract the crs of a \code{map}
}
\seealso{
Other mapMethods: \code{\link{area,map-method}},
  \code{\link{mapRm}},
  \code{\link{rasterToMatch,map-method}},
  \code{\link{rasters}}, \code{\link{studyAreaName}},
  \code{\link{studyArea}}
}
\concept{mapMethods}
