% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{customise_taxonomy_string}
\alias{customise_taxonomy_string}
\title{Customise a taxonomy string. Returns specified taxonomy levels. Assumes standard 7 ranks.}
\usage{
customise_taxonomy_string(
  x,
  taxa_levels = c("domain", "phylum", "class", "order", "family", "genus", "species")
)
}
\arguments{
\item{x}{taxonomy string}

\item{taxa_levels}{vector of taxa levels to return.}
}
\description{
Customise a taxonomy string. Returns specified taxonomy levels. Assumes standard 7 ranks.
}
\examples{
customise_taxonomy_string("d__Bacteria;p__Acidobacteriota", taxa_levels = c("phylum"))
}
