% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{darken}
\alias{darken}
\title{Darken a colour}
\usage{
darken(colour, factor = 1.4)
}
\arguments{
\item{colour}{colour name, hexadecimal string or positive integer \code{i} meaning \code{palette()[i]}}

\item{factor}{value to scale rgb colour vector by division}
}
\description{
Darken a colour
}
