% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{df2l}
\alias{df2l}
\title{Convert a dataframe to a list. Only first two columns are used.}
\usage{
df2l(mydataframe)
}
\arguments{
\item{mydataframe}{input dataframe}
}
\description{
Convert a dataframe to a list. Only first two columns are used.
}
