% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{df2m}
\alias{df2m}
\title{Convert a dataframe to a matrix. First column becomes row names}
\usage{
df2m(mydataframe)
}
\arguments{
\item{mydataframe}{input dataframe}
}
\description{
Convert a dataframe to a matrix. First column becomes row names
}
