% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{df_column_to_rownames}
\alias{df_column_to_rownames}
\title{Make row names of a dataframe the value of a defined column (default 1st) and then remove the column}
\usage{
df_column_to_rownames(mydataframe, rowname_col = 1)
}
\arguments{
\item{mydataframe}{input dataframe}

\item{rowname_col}{column to get row names from (default = 1)}
}
\description{
Make row names of a dataframe the value of a defined column (default 1st) and then remove the column
}
