% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS.R
\name{.setTilerPythonPath}
\alias{.setTilerPythonPath}
\title{Set python path for \code{tiler}}
\usage{
.setTilerPythonPath()
}
\description{
By default, \pkg{tiler} is configured to use python 2, which may not be available on
recent Linux distributions (e.g., Ubuntu 20.04+).
Thus, explicitly set tiler options to find the correct python path on their system.
}
\keyword{internal}
