% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{\%$\%}
\alias{\%$\%}
\title{Exposition pipe}
\arguments{
\item{lhs}{A list, environment, or a data.frame.}

\item{rhs}{An expression where the names in lhs is available.}
}
\value{
Result of rhs applied to one or several names of lhs.
}
\description{
Expose the names in `lhs` to the `rhs` expression.
Magrittr imported function, see details and examples in the magrittr package.
}
