% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{first_resolved_taxonomy}
\alias{first_resolved_taxonomy}
\title{Get the lowest taxonomy level (or 'first') that has been resolved from a taxonomy string}
\usage{
first_resolved_taxonomy(x)
}
\arguments{
\item{x}{taxonomy string}
}
\description{
Get the lowest taxonomy level (or 'first') that has been resolved from a taxonomy string
}
\examples{
first_resolved_taxonomy("d__Bacteria;p__Acidobacteriota;Unassigned)
first_resolved_taxonomy("d__Bacteria;p__Acidobacteriota;uncultured)
}
