% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{flattenCorrMatrix}
\alias{flattenCorrMatrix}
\title{Flatten a correlation and corresponding p-value}
\usage{
flattenCorrMatrix(cormat, pmat)
}
\arguments{
\item{cormat}{matrix of the correlation coefficients}

\item{pmat}{matrix of the correlation p-values}
}
\description{
Flatten a correlation and corresponding p-value
}
