% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{generate_correlation_network}
\alias{generate_correlation_network}
\title{Generate a correlation network}
\usage{
generate_correlation_network(
  cor_matrix.m,
  p_matrix.m = NULL,
  p_value_threshold = 0.05,
  cor_threshold = 0.5,
  filename = NULL,
  relabeller_function = NULL,
  node_size = 4,
  node_colour = "grey20",
  node_fill = "grey20",
  node_shape = 21,
  node_label_segment_colour = "grey",
  label_colour = "black",
  label_size = 1,
  plot_height = 10,
  plot_width = 10,
  plot_title = "",
  edge_width_min = 0.3,
  edge_width_max = 1,
  edge_alpha = 1,
  network_layout = "fr",
  exclude_to_from_df = NULL,
  myseed = NULL,
  edgetype = "link",
  show_p_label = F,
  file_type = "pdf"
)
}
\description{
Generate a correlation network
}
