% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_prevalence}
\alias{get_prevalence}
\title{Get prevalence of rows in matrix across columns}
\usage{
get_prevalence(matrix.m)
}
\arguments{
\item{matrix.m}{input matrix}
}
\description{
Get prevalence of rows in matrix across columns
}
