% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_pvalue_labels}
\alias{get_pvalue_labels}
\title{Create p-value label for input value}
\usage{
get_pvalue_labels(x)
}
\arguments{
\item{x}{integer/flout to create label for}
}
\value{
p-value label string, e.g. "***"
}
\description{
Create p-value label for input value
}
