% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_shape_list_from_metadata}
\alias{get_shape_list_from_metadata}
\title{Return list of shapes for variable values from metadata dataframe
assumes _shape column present}
\usage{
get_shape_list_from_metadata(metadata.df, variable)
}
\description{
Return list of shapes for variable values from metadata dataframe
assumes _shape column present
}
