% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largePatches.R
\name{LargePatches}
\alias{LargePatches}
\title{Calculate proportion of large patches in NRV}
\usage{
LargePatches(tsf, vtm, poly, labelColumn, id, ageClassCutOffs, ageClasses,
  sppEquivCol, sppEquivalencies_CA)
}
\arguments{
\item{tsf}{TODO: description needed}

\item{vtm}{TODO: description needed}

\item{poly}{TODO: description needed}

\item{labelColumn}{TODO: description needed}

\item{id}{TODO: description needed}

\item{ageClassCutOffs}{TODO: description needed}

\item{ageClasses}{TODO: description needed}
}
\description{
TODO: needs description
}
