% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leading.R
\name{leadingByStage}
\alias{leadingByStage}
\title{Calculate proportion of landscape occupied by each vegetation class}
\usage{
leadingByStage(tsf, vtm, polygonToSummarizeBy, ageClassCutOffs, ageClasses,
  objName = NULL, ...)
}
\arguments{
\item{tsf}{TODO}

\item{vtm}{TODO}

\item{polygonToSummarizeBy}{TODO}

\item{ageClassCutOffs}{TODO}

\item{ageClasses}{TODO}

\item{...}{TODO}
}
\value{
A \code{data.table} with proportion of the pixels in each vegetation class,
        for each given age class within each polygon.
}
\description{
This function is recursive.
If \code{polygonToSummarizeBy} is a \code{SpatialPolygon}, then the function
will enter once, and convert this to a fasterized version, and pass that into
the function replacing \code{polygonToSummarizeBy}.
It is also recursive of passed a vector of filenames for \code{tsf} and \code{vtm}.
}
