% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leading.R
\name{LeadingVegTypeByAgeClass}
\alias{LeadingVegTypeByAgeClass}
\title{Calculate proportion of landscape occupied by each vegetation class}
\usage{
LeadingVegTypeByAgeClass(tsf, vtm, poly, ageClassCutOffs, ageClasses)
}
\arguments{
\item{tsf}{A single filename, relative or absolute, pointing to a Time Since Fire raster.
Can be any format that \code{raster} can use.}

\item{vtm}{A single filename, relative or absolute, pointing to a Vegetation Type Map raster.
Can be any format that \code{raster} can use.}

\item{poly}{A single \code{SpatialPolygonsDataFrame} object or a factor \code{RasterLayer}.
This layer MUST have a column labelled \code{shinyLabel}}

\item{ageClassCutOffs}{A numeric vector with the endpoints for the \code{ageClasses}.
Should be \code{length(ageClasses) + 1}}

\item{ageClasses}{A character vector with labels for age classes to bin the \code{tsf} times,
e.g., \code{c("Young", "Immature", "Mature", "Old")}}
}
\value{
A \code{data.table} with proportion of the pixels in each vegetation class,
        for each given age class within each polygon.
}
\description{
This function is recursive.
If \code{poly} is a \code{SpatialPolygon}, then the function
will enter once, and convert this to a fasterized version, and pass that into
the function replacing \code{poly}.
It is also recursive of passed a vector of filenames for \code{tsf} and \code{vtm}.
}
