% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{m2df}
\alias{m2df}
\title{Convert a matrix to a dataframe. Rows become first column.}
\usage{
m2df(mymatrix, column_name = "Row_variable")
}
\arguments{
\item{column_name}{Name of the new column (defaults to "Row_variable")}
}
\description{
Convert a matrix to a dataframe. Rows become first column.
}
