% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-class.R
\docType{class}
\name{map-class}
\alias{map-class}
\alias{map}
\title{The \code{map} class}
\description{
Contains a common system for organzing vector and raster layers,
principally for use with \pkg{leaflet} and \pkg{shiny}.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{\code{data.table} with columns describing metadata of map objects in
\code{maps} slot.}

\item{\code{.xData}}{Named environment of map-type objects (e.g., \code{sf}, \code{Raster*},
\code{Spatial*}. Each entry may also be simply an environment, which
indicates where to find the object, i.e., via \code{get(layerName, envir = environment)}}

\item{\code{CRS}}{The common crs of all layers}

\item{\code{analyses}}{A \code{data.table} or \code{data.frame} of the types of analyses to perform.}

\item{\code{analysesData}}{A \code{data.table} or \code{data.frame} of the results of the analyses.}
}}

