% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbindlistAG.R
\name{rbindlistAG}
\alias{rbindlistAG}
\alias{rbindlist-analysisGroups}
\title{Utility functions for grouping analyses in a \code{map} object}
\usage{
rbindlistAG(map, functionName, analysisGroups)
}
\arguments{
\item{map}{Optional map object. If not provided, then one will be
created. If provided, then the present \code{object} or options passed to
prepInputs e.g., \code{url}, will be appended to this \code{map}}

\item{analysisGroups}{A character (length 1 currently), indicating which
analysis group (e.g., "analysisGroup1") should be used to rbindlist. Can
also specify "all" which will rbindlist all outputs}
}
\value{
A list of data.tables
}
\description{
Utility functions for grouping analyses in a \code{map} object
}
