% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{Map2}
\alias{Map2}
\title{Map and parallel::clusterMap together}
\usage{
Map2(..., cl = NULL)
}
\arguments{
\item{...}{passed to Map or clusterMap}

\item{cl}{passed to clusterMap}
}
\description{
This will send to Map or clusterMap, depending on whether cl is provided.
}
