% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{Map3}
\alias{Map3}
\title{Map/lapply/clusterMap/clusterApply all in one}
\usage{
Map3(fn, args, useCache)
}
\arguments{
\item{fn}{The function that will be run via Map/clusterMap/}

\item{args}{This MUST be a list of length 2, with first element
argsSingle and second element argsMulti. These latter are the
arguments that Map will cycle over.}
}
\description{
Usually run after \code{identifyVectorArgs}
}
