% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_NLP_PheWAS.R
\name{MAP_PheWAS_main}
\alias{MAP_PheWAS_main}
\title{MAP algorithm}
\usage{
MAP_PheWAS_main(dat.icd = NULL, dat.nlp = NULL, dat.note = NULL,
  nm.phe = NULL, nm.ID = NULL, nm.utl = NULL, p.icd = 0.001,
  n.icd = 100, p.nlp = 0.001, n.nlp = 100, wgt.icd.nlp = c("con.5",
  "con1", "sd")[1], yes.con = c(FALSE, TRUE)[1], yes.nlp = c(TRUE,
  FALSE)[1])
}
\arguments{
\item{dat.icd}{Data containing ICD codes counts and must have
nm.ID indicating the patient IDs.}

\item{dat.nlp}{Data containing NLP counts and must have
nm.ID indicating the patient IDs.}

\item{dat.note}{Data containing health utilization count and must have
nm.ID indicating the patient IDs.}

\item{nm.phe}{A vector of phenotypes whose results are desired.}

\item{nm.ID}{Variable name for the patient IDs that are shared by dat.icd,
dat.nlp, and dat.note.}

\item{nm.utl}{Variable name for the health utilization count in dat.note.}

\item{p.icd}{Threshold in percentage to check if ICD data is good to run MAP.}

\item{n.icd}{Threshold in count to check if ICD data is good to run MAP
(ICD has to pass both p.icd and n.icd threshold to be included ).}

\item{p.nlp}{Threshold in percentage to check if NLP data is good to run MAP.}

\item{n.nlp}{Threshold in count to check if NLP data is good to run MAP.
(NLP has to pass both p.nlp and n.nlp threshold to be included ).}

\item{wgt.icd.nlp}{Weighting scheme to create a new variable using ICD and NLP.}

\item{yes.con}{A logical variable indicating if concomitant is desired.}

\item{yes.nlp}{A logical variable indicating if patients having zero ICD, but nonzero
NLP are set aside.}
}
\value{
Returns a list with following objects:
\item{MAP}{A data.frame with each column being the probabilities for each phenotype.}
\item{cut.MAP}{A vector of cutoff values for the probabilities for each phenotype.}
}
\description{
Main function to perform MAP algorithm to calculate predicted probabilities of positive phenotype for each patient
based on NLP and ICD counts adjusted for healthcare utilization.
}
\examples{
## simulate data to test the algorithm
## see the vignette for more examples
ncase = 500 ; ncontrol = 2000
dd = c(rep(0,ncontrol), rep(1,ncase))
note = rpois(ncase+ncontrol, 20)
ICD = rpois(ncase+ncontrol, dd*3 + 0.1 + 0.1*log(note))
NLP = rpois(ncase+ncontrol, dd*3 + 0.1 + 0.1*log(note))
dat.icd = data.frame(ID=1:length(ICD), disease = ICD)
dat.nlp = data.frame(ID=1:length(ICD), disease = NLP)
dat.note = data.frame(ID=1:length(ICD), note = note)
nm.phe = 'disease' ; nm.ID = 'ID' ; nm.utl = 'note';
res = MAP_PheWAS_main(dat.icd = dat.icd, dat.nlp = dat.nlp, dat.note = dat.note,
nm.phe = nm.phe, nm.ID = nm.ID, nm.utl = nm.utl)
boxplot(res$MAP~dd)
}
\references{
Integrating natural language processing for high-throughput Multimodal
Automated Phenotyping (MAP) with Application to PheWAS. Katherine P. Liao, Jiehuan Sun,
Tianrun A. Cai, Nicholas Link, Chuan Hong, Jie Huang, Jennifer Huffman, Jessica Gronsbell, Yichi Zhang,
Yuk-Lam Ho, Jacqueline Honerlaw, Lauren Costa, Victor Castro, Vivian Gainer, Shawn Murphy,
Christopher J. O’Donnell, J. Michael Gaziano, Kelly Cho, Peter Szolovits, Isaac Kohane,
Sheng Yu, and Tianxi Cai with the VA Million Veteran Program
}
