% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-analysis.R
\name{mapAddAnalysis}
\alias{mapAddAnalysis}
\title{Add an analysis to a \code{map} object}
\usage{
mapAddAnalysis(map, functionName,
  useParallel = getOption("map.useParallel", FALSE), ...)
}
\arguments{
\item{map}{A \code{map} object}

\item{functionName}{The name of the analysis function to add}

\item{useParallel}{Logical indicating whether to use multiple threads.
Defaults to \code{getOption("map.useParallel", FALSE)}.}

\item{...}{Additional arguments TODO: description needed}
}
\description{
TODO: description neeeded
}
