% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{methods}
\name{mapRm}
\alias{mapRm}
\alias{mapRm.default}
\alias{crs,map-method}
\alias{crs}
\alias{studyAreaName}
\alias{studyAreaName.map}
\alias{studyArea}
\alias{studyArea.map}
\alias{rasters}
\alias{rasters.map}
\title{Map class methods}
\usage{
mapRm(map, layer, ask = TRUE, ...)

\method{mapRm}{default}(map = NULL, layer = NULL, ask = TRUE, ...)

\S4method{crs}{map}(x, ...)

studyAreaName(map, layer)

\method{studyAreaName}{map}(map, layer = 1)

studyArea(map, layer)

\method{studyArea}{map}(map, layer = 1)

rasters(map)

\method{rasters}{map}(map)
}
\description{
Tools for getting objects and metadata in and out of a \code{map} class.
}
