% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{rasters}
\alias{rasters}
\alias{rasters.map}
\alias{sp}
\alias{sp.map}
\alias{sf}
\alias{sf.map}
\alias{spatialPolygons}
\alias{spatialPoints}
\alias{maps}
\title{Extract rasters in the \code{map} object}
\usage{
rasters(map)

\method{rasters}{map}(map)

sp(map)

\method{sp}{map}(map)

sf(map)

\method{sf}{map}(map)

spatialPolygons(map)

spatialPoints(map)

maps(map, class = NULL, layerName = NULL)
}
\arguments{
\item{map}{A \code{map} class obj}

\item{class}{If supplied, this will be the class of objects returned. Default
is \code{NULL} which is "all", meaning all objects in the \code{map}
obj}
}
\value{
A list of maps (i.e., sp, raster, or sf objects) of class \code{class}
}
\description{
This will extract all objects in or pointed to within the \code{map}.
}
\seealso{
Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasterToMatch,map-method}},
  \code{\link{studyAreaName}}, \code{\link{studyArea}}

Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasterToMatch,map-method}},
  \code{\link{studyAreaName}}, \code{\link{studyArea}}

Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasterToMatch,map-method}},
  \code{\link{studyAreaName}}, \code{\link{studyArea}}

Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasterToMatch,map-method}},
  \code{\link{studyAreaName}}, \code{\link{studyArea}}

Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasterToMatch,map-method}},
  \code{\link{studyAreaName}}, \code{\link{studyArea}}

Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasterToMatch,map-method}},
  \code{\link{studyAreaName}}, \code{\link{studyArea}}
}
\concept{mapMethods}
