% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{metadata}
\alias{metadata}
\alias{metadata.map}
\alias{metadata.Map}
\alias{metadata.Raster}
\alias{metadata.SpatRaster}
\title{Extract metadata}
\usage{
metadata(x, ...)

\method{metadata}{map}(x, ...)

\method{metadata}{Map}(x, ...)

\method{metadata}{Raster}(x, ...)

\method{metadata}{SpatRaster}(x, ...)
}
\arguments{
\item{x}{A \code{map}, \code{Raster}, or \code{SpatRaster} object}

\item{...}{Additional arguments passed to other methods (not used)}
}
\description{
Extract metadata
}
