% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP.R
\docType{data}
\name{phecode.cuis.list}
\alias{phecode.cuis.list}
\title{MAP dictionary}
\format{
A list of 1866
}
\usage{
phecode.cuis.list
}
\description{
MAP dictionary that maps phecode to CUIs
}
\examples{
head(phecode.cuis.list)
tail(phecode.cuis.list)
}
\keyword{datasets}
