% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\value{
Result of rhs applied to lhs, see details in magrittr package.
}
\description{
Pipe an object forward into a function or call expression.
Magrittr imported function, see details and examples in the magrittr package.
}
