% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{plot_corrplot}
\alias{plot_corrplot}
\title{Generate a dot correlation plot generated by corrplot}
\usage{
plot_corrplot(
  correlation_matrix,
  p_value_matrix = NULL,
  plot_title = "",
  plot_title_size = 0.6,
  plot_height = 10,
  plot_width = 10,
  p_value_threshold = 0.05,
  label_size = 1,
  relabeller_function = NULL,
  insig = "blank",
  insig_pch = 4,
  insig_pch_cex = 1,
  insig_pch_col = "black",
  make_insig_na = F,
  to_exclude = NULL,
  method = "circle",
  outline = T,
  label_colour = "black",
  colour_label_size = 1,
  grid_colour = "black",
  pairs_to_na = NULL,
  order = "hclust",
  col = NULL,
  filename = NULL,
  file_type = "pdf",
  ...
)
}
\description{
Generate a dot correlation plot generated by corrplot
}
