% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-analysis.R
\name{mapAddPostHocAnalysis}
\alias{mapAddPostHocAnalysis}
\title{Add a post hoc analysis function to a \code{map} object}
\usage{
mapAddPostHocAnalysis(map, functionName, postHocAnalysisGroups = NULL,
  postHocAnalyses = "all", useParallel = getOption("map.useParallel"),
  ...)
}
\arguments{
\item{map}{Optional map object. If not provided, then one will be
created. If provided, then the present \code{object} or options passed to
prepInputs e.g., \code{url}, will be appended to this \code{map}}

\item{functionName}{A function that is designed for post hoc analysis of
map class objects, e.g., \code{rbindlistAG}}

\item{postHocAnalysisGroups}{Character string with one
\code{analysisGroups} i.e., \code{"analysisGroup1"} or \code{"analysisGroup2"}}

\item{postHocAnalyses}{Character vector with \code{"all"},
(which will do all analysisGroups) the default,
or 1 or more of the the \code{functionName}s that are in the analyses slot.}

\item{useParallel}{Logical. If \code{TRUE}, then if there is more than one
calculation to do at any stage, it will create and use a parallel
cluster via \code{makeOptimalCluster}}

\item{...}{Optional arguments to pass into \code{functionName}}
}
\description{
Add a post hoc analysis function to a \code{map} object
}
