% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{rasterToMatch}
\alias{rasterToMatch}
\alias{rasterToMatch.map}
\title{Extract the rasterToMatch(s) from a \code{map}}
\usage{
rasterToMatch(map, layerName, layer)

\method{rasterToMatch}{map}(map, layerName, layer = NA)
}
\arguments{
\item{map}{TODO: describe this}

\item{layerName}{TODO: describe this}

\item{layer}{TODO: describe this}
}
\description{
If \code{layer} is not provided and there is more than one \code{studyArea},
then this will extract the last one added.
}
\seealso{
Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasters}}, \code{\link{studyAreaName}},
  \code{\link{studyArea}}

Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasters}}, \code{\link{studyAreaName}},
  \code{\link{studyArea}}
}
\concept{mapMethods}
