% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove_from_vector}
\alias{remove_from_vector}
\title{Convenience function to remove item from a vector}
\usage{
remove_from_vector(vector, to_remove)
}
\arguments{
\item{to_remove}{items to remove}

\item{input}{vector}
}
\description{
Convenience function to remove item from a vector
}
