% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove_inf_na_matrix}
\alias{remove_inf_na_matrix}
\title{Remove inf and na values from a matrix}
\usage{
remove_inf_na_matrix(mymatrix)
}
\arguments{
\item{mymatrix}{input matrix}
}
\description{
Remove inf and na values from a matrix
}
