% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxPlots.R
\name{runHistsLargePatches}
\alias{runHistsLargePatches}
\title{Generate histograms for large patches}
\usage{
runHistsLargePatches(map, functionName, analysisGroups, dPath)
}
\arguments{
\item{map}{A \code{map} object.}

\item{functionName}{TODO: description needed}

\item{analysisGroups}{TODO: description needed}

\item{dPath}{Destination path for the resulting png files.}
}
\description{
TODO: description needed
}
