% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-analysis.R
\name{runMapAnalyses}
\alias{runMapAnalyses}
\title{\code{runMapAnalyses}}
\usage{
runMapAnalyses(
  map,
  purgeAnalyses = NULL,
  useParallel = getOption("map.useParallel", FALSE),
  ...
)
}
\arguments{
\item{map}{TODO}

\item{purgeAnalyses}{TODO}

\item{useParallel}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
TODO: description needed
}
