% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_data.R
\name{shape_list_from_dataframe}
\alias{shape_list_from_dataframe}
\title{Return list of shapes for variable values from metadata dataframe
assumes _shape column present}
\usage{
shape_list_from_dataframe(dataframe.df, variable)
}
\arguments{
\item{dataframe.df}{input dataframe}

\item{variable}{column (variable) to get shapes for}
}
\description{
Return list of shapes for variable values from metadata dataframe
assumes _shape column present
}
