% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{methods}
\name{show,map-method}
\alias{show,map-method}
\title{Show method for map class objects}
\usage{
\S4method{show}{map}(object)
}
\arguments{
\item{obj}{TODO: describe this}
}
\description{
Show method for map class objects
}
