% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{studyArea}
\alias{studyArea}
\alias{studyArea.map}
\title{Extract the studyArea(s) from a \code{map}}
\usage{
studyArea(map, layer)

\method{studyArea}{map}(map, layer = 1)
}
\description{
Extract the studyArea(s) from a \code{map}
}
\seealso{
Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasters}}, \code{\link{studyAreaName}}

Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasters}}, \code{\link{studyAreaName}}
}
\concept{mapMethods}
