% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{studyAreaName}
\alias{studyAreaName}
\alias{studyAreaName.map}
\title{Map class methods}
\usage{
studyAreaName(map, layerName, layer)

\method{studyAreaName}{map}(map, layerName, layer = 1)
}
\arguments{
\item{map}{TODO: document this}

\item{layerName}{TODO: document this}

\item{layer}{TODO: document this}
}
\description{
Tools for getting objects and metadata in and out of a \code{map} class.
}
\seealso{
Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasterToMatch}}, \code{\link{rasters}},
  \code{\link{studyArea}}

Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasterToMatch}}, \code{\link{rasters}},
  \code{\link{studyArea}}
}
\concept{mapMethods}
