% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{methods}
\name{area,map-method}
\alias{area,map-method}
\title{Calculate area of (named) objects the \code{map} obj}
\usage{
\S4method{area}{map}(x)
}
\arguments{
\item{x}{Raster* or SpatialPolygons object}
}
\description{
Calculate area of (named) objects the \code{map} obj
}
\seealso{
Other mapMethods: \code{\link{crs,map-method}},
  \code{\link{mapRm}},
  \code{\link{rasterToMatch,map-method}},
  \code{\link{rasters}}, \code{\link{studyAreaName}},
  \code{\link{studyArea}}
}
\concept{mapMethods}
