% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS.R
\name{areaAndPolyValue}
\alias{areaAndPolyValue}
\title{\code{areaAndPolyValue}}
\usage{
areaAndPolyValue(ras)
}
\arguments{
\item{ras}{A \code{Raster*} object}
}
\value{
list containing: \code{sizeInHa}, the area; and \code{polyID}, the polygon ID.
}
\description{
Determine the area of each zone in a raster. TODO: improve description
}
