% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMetadata.R
\name{buildMetadata}
\alias{buildMetadata}
\title{Build \code{map} obj metadata table}
\usage{
buildMetadata(metadata, isStudyArea, layerName, obj, columnNameForLabels,
  objHash, leaflet, envir, ...)
}
\description{
Build \code{map} obj metadata table
}
