% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMetadata.R
\name{buildMetadata}
\alias{buildMetadata}
\title{Build \code{map} obj metadata table}
\usage{
buildMetadata(metadata, isStudyArea, isRasterToMatch, layerName, obj,
  columnNameForLabels, objHash, leaflet, envir, ...)
}
\arguments{
\item{metadata}{TODO: description needed}

\item{isStudyArea}{TODO: description needed}

\item{isRasterToMatch}{Logical. Is this(these) layer(s) the \code{rasterToMatch} layers.
If \code{TRUE}, then this layer can be accessed by \code{rasterToMatch(map)}}

\item{layerName}{TODO: description needed}

\item{obj}{TODO: description needed}

\item{columnNameForLabels}{TODO: description needed}

\item{objHash}{TODO: description needed}

\item{leaflet}{TODO: description needed}

\item{envir}{TODO: description needed}

\item{...}{Additional arguments.}
}
\description{
Build \code{map} obj metadata table
}
