% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{crs,map-method}
\alias{crs,map-method}
\title{Extract the CRS of a \code{map}}
\usage{
\S4method{crs}{map}(x, ...)
}
\arguments{
\item{x}{Raster* or Spatial object }

\item{...}{additional arguments. None implemented}
}
\description{
Extract the CRS of a \code{map}
}
\seealso{
Other mapMethods: 
\code{\link{mapRm}()},
\code{\link{rasterToMatch,map-method}},
\code{\link{rasters}()},
\code{\link{studyArea}()},
\code{\link{studyAreaName}()}
}
\concept{mapMethods}
