% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMetadata.R
\name{.enforceColumnTypes}
\alias{.enforceColumnTypes}
\title{Ensure columns in metadata table are a particular type (class)}
\usage{
.enforceColumnTypes(metadata)
}
\arguments{
\item{metadata}{}
}
\value{
metadata \code{data.table} with corrected column types
}
\description{
Ensure columns in metadata table are a particular type (class)
}
\keyword{internal}
