% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS.R
\name{fasterize2}
\alias{fasterize2}
\title{\code{fasterize} following crop & reproject}
\usage{
fasterize2(emptyRaster, polygonToFasterize, field)
}
\arguments{
\item{emptyRaster}{An empty \code{RasterLayer} with \code{res}, \code{crs}, \code{extent} all
correct for to pass to \code{\link[=fasterize]{fasterize()}}}

\item{polygonToFasterize}{passed to \code{\link[=fasterize]{fasterize()}};
will be cropped first if \code{extent(emptyRaster) < extent(polygonToFasterize)}.}

\item{field}{passed to \code{\link[=fasterize]{fasterize()}}}
}
\description{
\code{fasterize} following crop & reproject
}
