% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leading.R
\name{fasterize2}
\alias{fasterize2}
\title{Fasterize with crop & spTransform first}
\usage{
fasterize2(emptyRaster, polygonToFasterize, field)
}
\arguments{
\item{emptyRaster}{An empty raster with res, crs, extent all
correct for to pass to \code{fasterize}}

\item{polygonToFasterize}{passed to \code{fasterize}, but it
will be cropped first if
\code{extent(emptyRaster) < extent(polygonToFasterize)}}

\item{field}{passed to fasterize}
}
\description{
Fasterize with crop & spTransform first
}
