% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS.R
\name{gdal_polygonizeR}
\alias{gdal_polygonizeR}
\title{Polygonize a raster}
\usage{
gdal_polygonizeR(
  x,
  outshape = NULL,
  gdalformat = "ESRI Shapefile",
  pypath = NULL,
  readpoly = TRUE,
  quiet = TRUE
)
}
\arguments{
\item{x}{a \code{RsaterLayer}, \code{SpatRaster}, or character giving the filepath to a raster.}

\item{outshape}{character giving the filepath for the output shapefile.}

\item{gdalformat}{GDAL driver to use. See \code{terra::gdal(drivers=TRUE)}.}

\item{pypath, quiet}{deprecated. maintained for backwards compatibility only (not used).}

\item{readpoly}{logical indicating whether the polygons object should be returned
(this was previously using \code{gdal_polygonize} on disk and required reading the file)}
}
\value{
if \code{readpoly = TRUE} (default), an \code{sf} polygons object; otherwise, \code{NULL}.
}
\description{
Polygonize a raster
}
