% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largePatches.R
\name{gdal_polygonizeR}
\alias{gdal_polygonizeR}
\title{Polygonize with GDAL}
\usage{
gdal_polygonizeR(x, outshape = NULL, gdalformat = "ESRI Shapefile",
  pypath = NULL, readpoly = TRUE, quiet = TRUE)
}
\description{
Copied from https://johnbaumgartner.wordpress.com/2012/07/26/getting-rasters-into-shape-from-r/
}
