% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTiles.R
\name{makeTiles}
\alias{makeTiles}
\title{Make tiles (pyramids) using \code{gdal2tiles}}
\usage{
makeTiles(tilePath, obj, overwrite = FALSE, ...)
}
\arguments{
\item{tilePath}{A director to write tiles}

\item{obj}{A raster object with or without file-backing}

\item{overwrite}{Logical. If \code{FALSE}, and the director exists,
then it will not overwrite any files.}

\item{...}{Arguments passed to \code{\link[reproducible:deprecated]{reproducible::projectInputs()}} (e.g., \code{useGDAL}).}
}
\description{
Make tiles (pyramids) using \code{gdal2tiles}
}
