% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTiles.R
\name{makeTiles}
\alias{makeTiles}
\alias{canMakeTiles}
\title{Make raster tiles (pyramids)}
\usage{
makeTiles(
  tilePath = getOption("map.tilePath", "tiles"),
  obj,
  overwrite = FALSE,
  ...
)

canMakeTiles()
}
\arguments{
\item{tilePath}{A directory to write tiles}

\item{obj}{A raster object with or without file-backing}

\item{overwrite}{Logical. If \code{FALSE}, and the director exists,
then it will not overwrite any files.}

\item{...}{Arguments passed to \code{\link[reproducible:deprecated]{reproducible::projectInputs()}} (e.g., \code{useGDAL}).}
}
\value{
\itemize{
\item \code{makeTiles()} is invoked for it's side-effect of creating tiles in \code{tilePath};
\item \code{canMakeTiles()} returns a logical indicating the availability of python and GDAL support;
}
}
\description{
If python with GDAL support is available, uses \code{gdal2tiles} via \code{\link[tiler:tile]{tiler::tile()}}
to create tiles for use with e.g. \code{leaflet}.
}
\note{
Requires python with GDAL support (see \code{\link[tiler:tiler_options]{tiler::tiler_options()}}).
Windows users should install 'OSGeo4W' (\url{https://trac.osgeo.org/osgeo4w/}).
}
