% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTiles.R
\name{makeTiles}
\alias{makeTiles}
\title{Make tiles (pyramids) using gdal2tiles}
\usage{
makeTiles(tilePath, obj, overwrite = FALSE, ...)
}
\arguments{
\item{tilePath}{A director to write tiles}

\item{obj}{A raster objects with or without file-backing}

\item{overwrite}{Logical. If \code{FALSE}, and the director exists,
then it will not overwrite any files.}
}
\description{
Make tiles (pyramids) using gdal2tiles
}
