% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTiles.R
\name{makeTiles}
\alias{makeTiles}
\title{Make tiles (pyramids) using \code{gdal2tiles}}
\usage{
makeTiles(tilePath, obj, overwrite = FALSE, ...)
}
\arguments{
\item{tilePath}{A director to write tiles}

\item{obj}{A raster objects with or without file-backing}

\item{overwrite}{Logical. If \code{FALSE}, and the director exists,
then it will not overwrite any files.}

\item{...}{Passed to \code{reproducible::projectInputs} e.g., \code{useGDAL}}
}
\description{
NOTE: by default, \pkg{tiler} is configured to use python 2, which may not be available on
recent Linux distributions (e.g., Ubuntu 20.04).
Thus, the user should explicitly set tiler options to find the correct python path on their
system, using e.g., \code{tiler::tiler_options(python = Sys.which("python3"))}.
}
