% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-class.R
\docType{class}
\name{map-class}
\alias{map-class}
\title{The \code{map} class}
\description{
Contains a common system for organizing geospatial vector and raster data,
principally for use with \pkg{leaflet} and \pkg{shiny}.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{\code{data.table} with columns describing metadata of objects in the \code{map}.}

\item{\code{.xData}}{Named environment of geospatial data objects (e.g., \code{sf}, \verb{Raster*}, \verb{Spatial*}).
Each entry may also be simply an environment, which indicates
where to find the object, i.e., via \code{get(layerName, envir = environment)}.}

\item{\code{CRS}}{The common CRS of all layers.}

\item{\code{paths}}{A named list of file paths. The default is a list of length 2 specifying
\code{dataPath} and \code{tilePath}.}

\item{\code{analyses}}{A \code{data.table} or \code{data.frame} of the types of analyses to perform.}

\item{\code{analysesData}}{A \code{data.table} or \code{data.frame} of the results of the analyses.}
}}

