% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-package.R
\docType{package}
\name{map-package}
\alias{map}
\alias{map-package}
\title{map: Defines a meta class of geographical objects, the 'map' class, and associated tools}
\description{
The 'map' class is a collection of geospatial objects (e.g., from the 'sp', 'raster', 'sf', and 'terra' packages), with a number of metadata additions to enable powerful methods, e.g., for leaflet, reproducible GIS etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://predictiveecology.github.io/map/}
  \item \url{https://github.com/PredictiveEcology/map}
  \item Report bugs at \url{https://github.com/PredictiveEcology/map/issues}
}

}
\author{
\strong{Maintainer}: Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})

Authors:
\itemize{
  \item Eliot J B McIntire \email{eliot.mcintire@nrcan-rncan.gc.ca} (\href{https://orcid.org/0000-0002-6914-8316}{ORCID})
}

Other contributors:
\itemize{
  \item His Majesty the King in Right of Canada, as represented by the Minister of Natural Resources Canada [copyright holder]
}

}
\keyword{internal}
