% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mapAdd}
\alias{mapAdd}
\alias{mapAdd.default}
\alias{mapAdd.SpatialPolygons}
\title{Append a spatial object to map}
\usage{
mapAdd(object, map, layerName, overwrite = FALSE, ...)

\method{mapAdd}{default}(object = NULL, map = NULL, layerName = NULL,
  overwrite = FALSE, sourceURL = NULL,
  columnNameForLabels = character(), leafletVisible = TRUE,
  studyArea = FALSE, ...)

\method{mapAdd}{SpatialPolygons}(object, map = NULL, layerName = NULL,
  overwrite = FALSE, sourceURL = NULL, columnNameForLabels = NULL,
  leafletVisible = TRUE, studyArea = NULL, ...)
}
\arguments{
\item{...}{passed to reproducible::postProcess and reproducible::projectInputs and
reproducible::fixErrors}
}
\description{
Append a spatial object to map
}
\examples{
library(sp)
library(raster)
coords <- structure(c(-122.98, -116.1, -99.2, -106, -122.98,
                      59.9, 65.73, 63.58, 54.79, 59.9),
                    .Dim = c(5L, 2L))
Sr1 <- Polygon(coords)
Srs1 <- Polygons(list(Sr1), "s1")
StudyArea <- SpatialPolygons(list(Srs1), 1L)
crs(StudyArea) <- "+init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"

ml <- new("map")
mapAdd(StudyArea, ml, studyArea = TRUE, layerName = "newPoly")

}
