% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mapAdd}
\alias{mapAdd}
\alias{mapAdd.default}
\alias{mapAdd.spatialObjects}
\title{Append a spatial object to map}
\usage{
mapAdd(object, map, layerName, overwrite = FALSE, ...)

\method{mapAdd}{default}(object = NULL, map = NULL, layerName = NULL,
  overwrite = FALSE, sourceURL = NULL,
  columnNameForLabels = character(), leaflet = TRUE,
  isStudyArea = FALSE, ...)

\method{mapAdd}{spatialObjects}(object, map = NULL, layerName = NULL,
  overwrite = FALSE, sourceURL = NULL, columnNameForLabels = NULL,
  leaflet = TRUE, isStudyArea = NULL, envir = NULL, ...)
}
\arguments{
\item{...}{passed to reproducible::postProcess and reproducible::projectInputs and
reproducible::fixErrors and reproducible::prepInputs}

\item{envir}{An optional environment. If supplied, then the object
will not be placed "into" the maps slot, rather the environment label will
be placed into the maps slot. Upon re}
}
\description{
Append a spatial object to map
}
\details{
If \code{isStudyArea = TRUE}, then several things will be triggered:

1. This layer will be added to metadata with \code{studyArea} set to
   \code{max(\code{studyArea(map)}) + 1}
2. update CRS slot to be the CRS of the study area
2.
}
\examples{
library(sp)
library(raster)
coords <- structure(c(-122.98, -116.1, -99.2, -106, -122.98,
                      59.9, 65.73, 63.58, 54.79, 59.9),
                    .Dim = c(5L, 2L))
Sr1 <- Polygon(coords)
Srs1 <- Polygons(list(Sr1), "s1")
StudyArea <- SpatialPolygons(list(Srs1), 1L)
crs(StudyArea) <- "+init=epsg:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"

ml <- new("map")
ml <- mapAdd(StudyArea, ml, isStudyArea = TRUE, layerName = "newPoly")

if (requireNamespace("SpaDES.tools")) {
  smallStudyArea <- SpaDES.tools::randomPolygon(studyArea(ml), 1e2)
  ml <- mapAdd(smallStudyArea, ml, isStudyArea = TRUE, filename2 = NULL,
               envir = .GlobalEnv) # adds a second studyArea within 1st
}

}
