% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mapRm}
\alias{mapRm}
\alias{mapRm.default}
\title{Remove objects from a \code{map}}
\usage{
mapRm(map, layer, ask = TRUE, ...)

\method{mapRm}{default}(map = NULL, layer = NULL, ask = TRUE, ...)
}
\arguments{
\item{map}{TODO: document this}

\item{layer}{TODO: document this}

\item{ask}{TODO: document this}

\item{...}{TODO: document this}
}
\description{
Remove objects from a \code{map}
}
\examples{
if (require("SpaDES.tools", quietly = TRUE)) {
  p <- terra::vect(cbind(-120, 60), crs = "epsg:4326") |>
    SpaDES.tools::randomPolygon(area = 1e5) |>
    sf::st_as_sf() |>
    sf::as_Spatial()
  m <- mapAdd(p, layerName = "p")
  m

  m <- mapRm(m, "p")
  m
}
}
\seealso{
Other mapMethods: 
\code{\link{area,map-method}},
\code{\link{crs,map-method}},
\code{\link{rasterToMatch,map-method}},
\code{\link{rasters}()},
\code{\link{studyArea}()},
\code{\link{studyAreaName}()}

Other mapMethods: 
\code{\link{area,map-method}},
\code{\link{crs,map-method}},
\code{\link{rasterToMatch,map-method}},
\code{\link{rasters}()},
\code{\link{studyArea}()},
\code{\link{studyAreaName}()}
}
\concept{mapMethods}
