% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mapRm}
\alias{mapRm}
\alias{mapRm.default}
\title{Remove objects from a \code{map}}
\usage{
mapRm(map, layer, ask = TRUE, ...)

\method{mapRm}{default}(map = NULL, layer = NULL, ask = TRUE, ...)
}
\arguments{
\item{map}{TODO: document this}

\item{layer}{TODO: document this}

\item{ask}{TODO: document this}

\item{...}{TODO: document this}
}
\description{
Remove objects from a \code{map}
}
\seealso{
Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}},
  \code{\link{rasterToMatch}}, \code{\link{rasters}},
  \code{\link{studyAreaName}}, \code{\link{studyArea}}

Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}},
  \code{\link{rasterToMatch}}, \code{\link{rasters}},
  \code{\link{studyAreaName}}, \code{\link{studyArea}}
}
\concept{mapMethods}
