% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mapRm}
\alias{mapRm}
\alias{mapRm.default}
\title{Remove objects from a \code{map}}
\usage{
mapRm(map, layer, ask = TRUE, ...)

\method{mapRm}{default}(map = NULL, layer = NULL, ask = TRUE, ...)
}
\arguments{
\item{map}{TODO: document this}

\item{layer}{TODO: document this}

\item{ask}{TODO: document this}

\item{...}{TODO: document this}
}
\description{
Remove objects from a \code{map}
}
\examples{
if (require("SpaDES.tools")) {
  library(sp)
  longLatCRS <- CRS(paste("+init=epsg:4326 +proj=longlat +datum=WGS84",
                          "+no_defs +ellps=WGS84 +towgs84=0,0,0"))
  p <- randomPolygon(SpatialPoints(cbind(-120, 60), proj4string = longLatCRS),
       area = 1e5)
  m <- mapAdd(p, layerName = "p")
  mapRm(m, "p")
}
}
\seealso{
Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}},
  \code{\link{rasterToMatch,map-method}},
  \code{\link{rasters}}, \code{\link{studyAreaName}},
  \code{\link{studyArea}}

Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}},
  \code{\link{rasterToMatch,map-method}},
  \code{\link{rasters}}, \code{\link{studyAreaName}},
  \code{\link{studyArea}}
}
\concept{mapMethods}
