% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{rasters}
\alias{rasters}
\alias{rasters.map}
\alias{sp}
\alias{sp.map}
\alias{sf}
\alias{sf.map}
\alias{spatialPolygons}
\alias{spatialPolygons.map}
\alias{spatialPoints}
\alias{spatialPoints.map}
\alias{spatRasters}
\alias{spatRasters.map}
\alias{spatVectors}
\alias{spatVectors.map}
\alias{maps}
\title{Extract objects of specific classes from a \code{map} object}
\usage{
rasters(map, ...)

\method{rasters}{map}(map, ...)

sp(map, ...)

\method{sp}{map}(map, ...)

sf(map, ...)

\method{sf}{map}(map, ...)

spatialPolygons(map, ...)

\method{spatialPolygons}{map}(map, ...)

spatialPoints(map, ...)

\method{spatialPoints}{map}(map, ...)

spatRasters(map, ...)

\method{spatRasters}{map}(map, ...)

spatVectors(map, ...)

\method{spatVectors}{map}(map, ...)

maps(map, class = NULL, layerName = NULL)
}
\arguments{
\item{map}{A \code{map} object}

\item{...}{Additional arguments passed to other methods (not used)}

\item{class}{If supplied, this will be the class of objects returned.
Default is \code{NULL} which is "all", meaning all objects in the \code{map} object.}

\item{layerName}{character string giving the name(s) of layer(s) to extract.}
}
\value{
A list of geospatial objects of class \code{class}
}
\description{
\itemize{
\item \code{rasters()} extracts \code{RasterLayer} objects;
\item \code{sf()} extracts \code{sf} objects;
\item \code{sp()} extracts \code{Spatial} objects;
\item \code{spatialPolygons()} extracts \code{SpatialPolygons} objects;
\item \code{spatialPoints()} extracts \code{SpatialPoints} objects;
\item \code{spatRasters()} extracts \code{SpatRaster} objects;
\item \code{spatVectors()} extracts \code{SpatVector} objects;
}

This will extract all objects in or pointed to within the \code{map}.
}
\seealso{
Other mapMethods: 
\code{\link{crs,map-method}},
\code{\link{mapRm}()},
\code{\link{rasterToMatch,map-method}},
\code{\link{studyAreaName}()},
\code{\link{studyArea}()}

Other mapMethods: 
\code{\link{crs,map-method}},
\code{\link{mapRm}()},
\code{\link{rasterToMatch,map-method}},
\code{\link{studyAreaName}()},
\code{\link{studyArea}()}

Other mapMethods: 
\code{\link{crs,map-method}},
\code{\link{mapRm}()},
\code{\link{rasterToMatch,map-method}},
\code{\link{studyAreaName}()},
\code{\link{studyArea}()}

Other mapMethods: 
\code{\link{crs,map-method}},
\code{\link{mapRm}()},
\code{\link{rasterToMatch,map-method}},
\code{\link{studyAreaName}()},
\code{\link{studyArea}()}

Other mapMethods: 
\code{\link{crs,map-method}},
\code{\link{mapRm}()},
\code{\link{rasterToMatch,map-method}},
\code{\link{studyAreaName}()},
\code{\link{studyArea}()}

Other mapMethods: 
\code{\link{crs,map-method}},
\code{\link{mapRm}()},
\code{\link{rasterToMatch,map-method}},
\code{\link{studyAreaName}()},
\code{\link{studyArea}()}

Other mapMethods: 
\code{\link{crs,map-method}},
\code{\link{mapRm}()},
\code{\link{rasterToMatch,map-method}},
\code{\link{studyAreaName}()},
\code{\link{studyArea}()}

Other mapMethods: 
\code{\link{crs,map-method}},
\code{\link{mapRm}()},
\code{\link{rasterToMatch,map-method}},
\code{\link{studyAreaName}()},
\code{\link{studyArea}()}

Other mapMethods: 
\code{\link{crs,map-method}},
\code{\link{mapRm}()},
\code{\link{rasterToMatch,map-method}},
\code{\link{studyAreaName}()},
\code{\link{studyArea}()}

Other mapMethods: 
\code{\link{crs,map-method}},
\code{\link{mapRm}()},
\code{\link{rasterToMatch,map-method}},
\code{\link{studyAreaName}()},
\code{\link{studyArea}()}

Other mapMethods: 
\code{\link{crs,map-method}},
\code{\link{mapRm}()},
\code{\link{rasterToMatch,map-method}},
\code{\link{studyAreaName}()},
\code{\link{studyArea}()}

Other mapMethods: 
\code{\link{crs,map-method}},
\code{\link{mapRm}()},
\code{\link{rasterToMatch,map-method}},
\code{\link{studyAreaName}()},
\code{\link{studyArea}()}
}
\concept{mapMethods}
