% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{metadata}
\alias{metadata}
\alias{metadata.Raster}
\alias{metadata.map}
\title{Extract the metadata object}
\usage{
metadata(x)

\method{metadata}{Raster}(x)

\method{metadata}{map}(x)
}
\arguments{
\item{x}{TODO: description needed}
}
\description{
Methods for specific classes exist.
}
