% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{rasterToMatch,map-method}
\alias{rasterToMatch,map-method}
\title{Extract the \code{rasterToMatch}(s) from a \code{map} object}
\usage{
\S4method{rasterToMatch}{map}(x, layer = 1)
}
\arguments{
\item{x}{a \code{map} object}

\item{layer}{an integer identifying the index of the \code{rasterToMatch} to extract}
}
\description{
If \code{layer} is not provided and there is more than one \code{studyArea},
then this will extract the last one added.
}
\seealso{
Other mapMethods: 
\code{\link{crs,map-method}},
\code{\link{mapRm}()},
\code{\link{rasters}()},
\code{\link{studyAreaName}()},
\code{\link{studyArea}()}
}
\concept{mapMethods}
