% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{rasterToMatch,map-method}
\alias{rasterToMatch,map-method}
\title{Extract the rasterToMatch(s) from a \code{x}}
\usage{
\S4method{rasterToMatch}{map}(x, layer = 1)
}
\arguments{
\item{x}{TODO: describe this}

\item{layer}{TODO: describe this}
}
\description{
If \code{layer} is not provided and there is more than one \code{studyArea},
then this will extract the last one added.
}
\seealso{
Other mapMethods: 
\code{\link{area,map-method}},
\code{\link{crs,map-method}},
\code{\link{mapRm}()},
\code{\link{rasters}()},
\code{\link{studyArea}()},
\code{\link{studyAreaName}()}
}
\concept{mapMethods}
