% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS.R
\name{.rasterToMemory}
\alias{.rasterToMemory}
\alias{.rasterToMemory.Raster}
\alias{.rasterToMemory.SpatRaster}
\title{\code{.rasterToMemory}}
\usage{
.rasterToMemory(x, ...)

\method{.rasterToMemory}{Raster}(x, ...)

\method{.rasterToMemory}{SpatRaster}(x, ...)
}
\arguments{
\item{x}{A \code{Raster} or \code{SpatRaster} object}

\item{...}{Additional arguments passed to raster read function}
}
\description{
\code{.rasterToMemory}
}
