% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS.R
\name{.rasterToMemory}
\alias{.rasterToMemory}
\title{\code{.rasterToMemory}}
\usage{
.rasterToMemory(x, ...)
}
\arguments{
\item{x}{A \verb{Raster*} object}

\item{...}{Additional arguments passed to \code{raster}}
}
\description{
\code{.rasterToMemory}
}
