% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{studyArea}
\alias{studyArea}
\alias{studyArea.map}
\alias{studyArea.default}
\title{Extract the studyArea(s) from a \code{map}}
\usage{
studyArea(map, layer)

\method{studyArea}{map}(map, layer = NA)

\method{studyArea}{default}(map, layer = NA)
}
\description{
If \code{layer} is not provided and there is more than one \code{studyArea},
then this will extract the last one added.
}
\seealso{
Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasterToMatch}}, \code{\link{rasters}},
  \code{\link{studyAreaName}}

Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasterToMatch}}, \code{\link{rasters}},
  \code{\link{studyAreaName}}

Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasterToMatch}}, \code{\link{rasters}},
  \code{\link{studyAreaName}}
}
\concept{mapMethods}
