% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{methods}
\name{studyArea}
\alias{studyArea}
\alias{studyArea,ANY-method}
\alias{studyArea,map-method}
\alias{studyArea<-,map-method}
\title{Extract the studyArea(s) from a \code{map}}
\usage{
studyArea(map, layer = NA)

\S4method{studyArea}{ANY}(map, layer = NA)

\S4method{studyArea}{map}(map, layer = NA)

\S4method{studyArea}{map}(map) <- value
}
\arguments{
\item{map}{TODO: document this}

\item{layer}{TODO: document this}

\item{layerName}{TODO: document this}
}
\description{
If \code{layer} is not provided and there is more than one \code{studyArea},
then this will extract the last one added.
}
\seealso{
Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasterToMatch,map-method}},
  \code{\link{rasters}}, \code{\link{studyAreaName}}

Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasterToMatch,map-method}},
  \code{\link{rasters}}, \code{\link{studyAreaName}}

Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasterToMatch,map-method}},
  \code{\link{rasters}}, \code{\link{studyAreaName}}

Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasterToMatch,map-method}},
  \code{\link{rasters}}, \code{\link{studyAreaName}}
}
\concept{mapMethods}
