% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{methods}
\name{studyArea}
\alias{studyArea}
\alias{studyArea,ANY-method}
\alias{studyArea,map-method}
\alias{studyArea<-}
\alias{studyArea<-,map-method}
\title{Extract the studyArea(s) from a \code{map}}
\usage{
studyArea(map, layer = NA, sorted = FALSE)

\S4method{studyArea}{ANY}(map, layer = NA, sorted = FALSE)

\S4method{studyArea}{map}(map, layer = NA, sorted = FALSE)

studyArea(map, layer = NA) <- value

\S4method{studyArea}{map}(map, layer = NA) <- value
}
\arguments{
\item{map}{TODO: document this}

\item{layer}{TODO: document this}

\item{sorted}{Logical. Should the numeric \code{layer} be referring to
geographic area of the \code{area} or the order that
the \code{studyArea} were placed into map object}

\item{value}{The value to assign to the object.}
}
\description{
If \code{layer} is not provided and there is more than one \code{studyArea},
then this will extract the last one added.
}
\seealso{
Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasterToMatch,map-method}},
  \code{\link{rasters}}, \code{\link{studyAreaName}}

Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasterToMatch,map-method}},
  \code{\link{rasters}}, \code{\link{studyAreaName}}

Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasterToMatch,map-method}},
  \code{\link{rasters}}, \code{\link{studyAreaName}}

Other mapMethods: \code{\link{area,map-method}},
  \code{\link{crs,map-method}}, \code{\link{mapRm}},
  \code{\link{rasterToMatch,map-method}},
  \code{\link{rasters}}, \code{\link{studyAreaName}}
}
\concept{mapMethods}
