% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{studyAreaName}
\alias{studyAreaName}
\alias{studyAreaName.map}
\title{Map class methods}
\usage{
studyAreaName(map, layer)

\method{studyAreaName}{map}(map, layer = 1)
}
\description{
Tools for getting objects and metadata in and out of a \code{map} class.
}
\seealso{
Other mapMethods: \code{\link{crs,map-method}},
  \code{\link{mapRm}}, \code{\link{rasters}},
  \code{\link{studyArea}}

Other mapMethods: \code{\link{crs,map-method}},
  \code{\link{mapRm}}, \code{\link{rasters}},
  \code{\link{studyArea}}
}
\concept{mapMethods}
